refactor(desktop): split deep-link tests
Keep the production deep-link module under the desktop file-size ratchet while preserving its test coverage unchanged.
